Rebase onto 7.0.0.0-R
This is a combination of 4 commits.
Do not dispose the titleImageLabel to resolve "Widget is disposed" error when opening the repository dialog
Comment-out the use of an unimplemented SWT API "gc.copyArea"
Fix the issue that menuitems cannot be pushed on webSpoon unless their type is checkbox
Disable the use of CMD key for shortcuts. At least, this should not be determined in the server-side