Rebase onto 8.2.0.0-342
This is a combination of the following commits.
Do not dispose the titleImageLabel to resolve "Widget is disposed" error when opening the repository dialog
Comment-out the use of an unimplemented SWT API "gc.copyArea"
Fix the issue that menuitems cannot be pushed on webSpoon unless their type is checkbox
Disable the use of CMD key for shortcuts. At least, this should not be determined in the server-side
Revert "Do not dispose the titleImageLabel to resolve "Widget is disposed" error when opening the repository dialog"
This reverts commit cd6fb78419235e717f3b437c00b192ca5c65c176.
Change the version of pentaho-xul-swt
Restore the image of checkboxes in XulTree